Add unit test for credential search.

+++ db/store/test/db_test.go
@@ -171,7 +171,7 @@ func validateCreatedTS(t *testing.T, cursor uint64, ts *time.Time) {
 		t.Errorf("Timestamp not in threshold %v", ts)
 	}
 	created := model.TimeToCursor(ts)
-	diff := uint64(math.Abs(float64(cursor - created)))
+	diff := uint64(math.Abs(float64(cursor) - float64(created)))
 	if diff > 1 {
 		t.Errorf("Cursor mismatch %v %v (%v)", cursor, created, diff)
 	}
